Introduce useUpdateAssessmentFiles, useDeleteAssessmentFile
