+++ src/client/store/ui/assessmentFiles/index.ts
@@ -1,3 +1,3 @@
-export { useAssessmentFiles, useGetAssessmentFiles } from './hooks'
+export { useAssessmentFiles, useGetAssessmentFiles, useUpdateAssessmentFiles } from './hooks'
 export { AssessmentFilesActions } from './slice'
 export type { AssessmentFilesState } from './stateType'
